Delete alert make beautiful

+++ src/components/DeleteButton.js
@@ -1,41 +1,51 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import Swal from "sweetalert2"
 
 function DeleteButton({ username }) {
   const [confirmed, setConfirmed] = useState(false);
   const navigate = useNavigate();
-  const handleDelete = () => {
-    let url =  "";
-    if(localStorage.getItem("role") ==="user"){
-      url = `http://127.0.0.1:5000/student/${username}`
-    } else{
-      url = `http://127.0.0.1:5000/teachers/${username}`
-    }
-    const options = {
-      method: 'DELETE',
-      headers: {
-        'Authorization': 'Basic ' + btoa(localStorage.getItem('username') + ':' + localStorage.getItem('password'))
-      }
-    };
-    if (confirmed) {
-      fetch(url, options)
-        .then(response => {
-          if (!response.ok) {
-            throw new Error('Network response was not ok');
-          }
-          navigate('/');
-          localStorage.clear();
-        })
-    }
-  }
 
   const handleClick = () => {
-    if (confirmed) {
-      handleDelete();
-    } else {
-      alert("Are you sure you want to delete this user?");
-      setConfirmed(true);
-    }
+      Swal.fire({
+        title: 'Are you sure you want to delete this account?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#32CD32',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes'
+      }).then((result) => {
+        if (result.isConfirmed) 
+        {
+          Swal.fire(
+            'Deleted!',
+            'Your account has been deleted',
+            'success'
+          ).then(() => {
+            let url =  "";
+            if(localStorage.getItem("role") ==="user"){
+              url = `http://127.0.0.1:5000/student/${username}`
+            } else{
+              url = `http://127.0.0.1:5000/teachers/${username}`
+            }
+            const options = {
+              method: 'DELETE',
+              headers: {
+                'Authorization': 'Basic ' + btoa(localStorage.getItem('username') + ':' + localStorage.getItem('password'))
+              }
+            };
+
+            fetch(url, options)
+              .then(response => {
+                if (!response.ok) {
+                  throw new Error('Network response was not ok');
+                }
+                localStorage.clear();
+                window.location.reload()
+              })
+          });
+        }
+      })  
   }
 
   return (
